% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{epiDat}
\alias{epiDat}
\title{Simulated Epidemiological Dataset.}
\format{
A data frame with 10000 rows and 3 variables:
\describe{
\item{death}{the death status of a subject (0 = alive, 1 = dead)}
\item{disease}{the disease status of a subject (0 = undiseased, 1 = diseased)}
\item{age}{the age of a subject, in years}
}
}
\usage{
epiDat
}
\description{
A simulated epidemiological dataset where each observation contains information
on a subject's \code{death} status, \code{disease} status, and \code{age}.
}
\keyword{datasets}
