% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_logistic_model.R
\name{fit_logistic_model}
\alias{fit_logistic_model}
\title{Fit a Logistic Regression Model}
\usage{
fit_logistic_model(formula, data)
}
\arguments{
\item{formula}{The formula for the model.}

\item{data}{The data that will be used to fit the model.}
}
\value{
The beta coefficients and the variance-covariance matrix for the model fit.
}
\description{
A helper function used to fit a logistic regression model.
}
