% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_margin_and_jacobian_row.R
\name{get_margin_and_jacobian_row}
\alias{get_margin_and_jacobian_row}
\title{Get an Estimated Predictive Margin and a Row of the Jacobian Matrix for a Logistic Regression Model}
\usage{
get_margin_and_jacobian_row(model_data, betas, main_effect, value, over)
}
\arguments{
\item{model_data}{The data used to fit the Logistic Regression model}

\item{betas}{The beta coefficients obtained from fitting the Logistic Regression model}

\item{main_effect}{The name of the variable corresponding to the main effect in your model.}

\item{value}{0 or 1, corresponding the value of \code{main_effect} at this margin.}

\item{over}{A Boolean variable.  If \code{TRUE}, compute the equivalent of the Stata command \code{margins, over(var_name)}.
If \code{FALSE}, compute the equivalent of \code{margins var_name}.}
}
\value{
The mean of the predicted probabilities and the corresponding row of the Jacobian matrix.
}
\description{
Helper function for \code{get_margins_results}.
}
