% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_margins.R
\name{get_margins}
\alias{get_margins}
\title{Get Marginal Probabilities for a Logistic Regression Model}
\usage{
get_margins(model_data, betas, sigma, main_effect, over)
}
\arguments{
\item{model_data}{The data used to fit the Logistic Regression model}

\item{betas}{The beta coefficients obtained from fitting the Logistic Regression model}

\item{sigma}{The variance-covariance matrix obtained from fitting the Logistic Regression model}

\item{main_effect}{The name of the variable corresponding to the main effect in your model.}

\item{over}{A Boolean variable.  If \code{TRUE}, compute the equivalent of the Stata command \code{margins, over(var_name)}.
If \code{FALSE}, compute the equivalent of \code{margins var_name}.}
}
\value{
The marginal probabilities, their standard errors, and their variance-covariance matrix.
}
\description{
For a logistic regression model that includes one main effect
and zero or more adjustors, obtain marginal probalities for
the outcome variable with respect to the main effect.
This reproduces the behavior of running \code{margins x} after
fitting the logistic regression model in Stata.
}
