% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_or_from_beta.R
\name{get_or_from_beta}
\alias{get_or_from_beta}
\title{Get the Odds Ratio using the Beta Estimate}
\usage{
get_or_from_beta(betas, sigma, main_effect)
}
\arguments{
\item{betas}{The beta coefficients obtained from fitting the Logistic Regression model}

\item{sigma}{The variance-covariance matrix obtained from fitting the Logistic Regression model}

\item{main_effect}{The variable name of the main effect in your model.}
}
\description{
If \code{beta_mar_var} is the Beta Estimate for the marginal variable \code{mar_var}
obtained after fitting a Logistic regression model, then this function returns
the Odd Ratio obatined by exponentiating this Beta Estimate: \code{epx(beta_mar_var)}.
For reference, this is the the Odds Ratio that would be obtained
by using the \code{logistic} command in Stata( \code{logistic outcome mar_var ...}).
}
