% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logisticMargins.R
\name{logisticMargins}
\alias{logisticMargins}
\title{Perform marginal analysis for a logistic regression model}
\usage{
logisticMargins(formula, main_effect, over, data)
}
\arguments{
\item{formula}{The formula for the logistic regression model}

\item{main_effect}{The variable name of the main effect in your model.}

\item{over}{A Boolean variable.  If \code{TRUE}, compute the equivalent of the Stata command \code{margins, over(var_name)}.
If \code{FALSE}, compute the equivalent of \code{margins var_name}.}

\item{data}{The data that will be used to fit the logistic regression model.}
}
\value{
A \code{data.frame} containing the predictive margins, a risk ratio, a relative risk, an odds ratio, and their standard errors.
}
\description{
For a logistic regression model with a binary main effect, obtain predicitive margins
along with their corresponding risk ratio, relative risk, odds ratio, and standard errors.
}
